Update index.html (#36)
The sentence "This website is presently under built" is not grammatically correct and is unclear in meaning.

The phrase "under built" is not a standard English expression. A possible alternative phrase that could be used is "under construction", which means that the website is being built or developed at the moment. So, a better version of the sentence could be "This website is presently under construction."

It's also worth noting that "presently" may be confusing in this context, as it can mean either "at the present time" or "soon". To avoid ambiguity, it may be better to use "currently" instead, as in "This website is currently under construction."